+++ src/shared/common-ui/page-nation/index.ts
@@ -1,0 +1,1 @@
+export * from "./page-nation";
