+++ src/shared/hooks/index.ts
@@ -1,1 +1,2 @@
 export * from "./use-localstorage";
+export * from "./use-page-nation";
